+++ docs/source/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Latest
+## 0.6.5 (2023-08-16)
 - Implemented automatic versioning of tasks by setting task version to [](#AUTO_VERSION).
   This feature is currently only supported by Polars [`LazyFrame`](inv:pl#reference/lazyframe/index) and by Pandas.
 - Added [](#kroki_url) config option.